Refactored categories, so each user can create their own category, opposed to choosing a predfined category

+++ src/Controller/ExpenseCategoryController.php
@@ -3,6 +3,8 @@
 namespace App\Controller;
 
 use App\Entity\ExpensesCategories;
+use App\Security\JwtAuthenticator;
+use JMS\Serializer\SerializationContext;
 use JMS\Serializer\SerializerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
@@ -10,6 +12,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Exception\HttpException;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
@@ -21,13 +24,21 @@ class ExpenseCategoryController extends Controller
 
     private $validator;
 
+    private $authenticator;
+
+    private $userProvider;
+
     public function __construct(
         SerializerInterface $serializer,
-        ValidatorInterface $validator
+        ValidatorInterface $validator,
+        JwtAuthenticator $authenticator,
+        UserProviderInterface $userProvider
     )
     {
         $this->serializer = $serializer;
         $this->validator = $validator;
+        $this->authenticator = $authenticator;
+        $this->userProvider = $userProvider;
     }
 
 	/**
@@ -38,22 +49,32 @@ class ExpenseCategoryController extends Controller
      */
     public function index(Request $request): JsonResponse
 	{
+        $user = $this->authenticator->getUser(
+            $this->authenticator->getCredentials($request),
+            $this->userProvider
+        );
+
         $page = $request->query->get('page');
         $limit = $request->query->get('limit');
 
         $repository = $this->getDoctrine()->getRepository(ExpensesCategories::class);
 
         if(!$page || !is_numeric($page) || !$limit || !is_numeric($limit)) {
-            $results = $repository->getAll();
+            $results = $repository->getAll($user->getId());
         } else {
-            $results = $repository->getAll((int)$page, (int)$limit);
+            $results = $repository->getAll($user->getId(), (int)$page, (int)$limit);
         }
 
         if (!$results) {
-            return new JsonResponse(['success' => true, 'data' => []]);
+            return new JsonResponse(['data' => []]);
         }
         
-        $categories = $this->serializer->serialize(['success' => true, 'data' => $results], 'json');
+        $categories = $this->serializer->serialize([
+            'data' => $results],
+            'json',
+            SerializationContext::create()
+                ->setGroups(['Default'])
+        );
 
         return new JsonResponse($categories, 200, [], true);
 	}
@@ -61,12 +82,18 @@ class ExpenseCategoryController extends Controller
 	/**
 	 * @Route("/expense-categories/{id}", name="get_expense_category")
      * @Method("GET")
+     * @param Request $request
      * @param string $id
      * @return JsonResponse
      * @throws HttpException
      */
-    public function getCategory(string $id): JsonResponse
+    public function getCategory(Request $request, string $id): JsonResponse
     {
+        $user = $this->authenticator->getUser(
+            $this->authenticator->getCredentials($request),
+            $this->userProvider
+        );
+
         $em = $this->getDoctrine()->getManager();
         $result = $em->getRepository(ExpensesCategories::class)->find($id);
 
@@ -74,7 +101,16 @@ class ExpenseCategoryController extends Controller
             throw new HttpException(404,'Resource does not exist.');
         }
 
-        $category = $this->serializer->serialize(['success' => true, 'data' => $result], 'json');
+        if($user->getId() != $result->getUser()->getId()) {
+            throw new HttpException(400,'You do not have permission to view this resource.');
+        }
+
+        $category = $this->serializer->serialize([
+            'data' => $result],
+            'json',
+            SerializationContext::create()
+                ->setGroups(['Default'])
+        );
 
 		return new JsonResponse($category, 200, [], true);
 	}
@@ -101,10 +137,16 @@ class ExpenseCategoryController extends Controller
             throw new HttpException(400, 'You must provide category property');
         }
 
+        $user = $this->authenticator->getUser(
+            $this->authenticator->getCredentials($request),
+            $this->userProvider
+        );
+
         $now = new \DateTime('now', new \DateTimeZone('Europe/Ljubljana'));
 
         $expenseCategory = new ExpensesCategories();
         $expenseCategory->setCategory($data->getCategory());
+        $expenseCategory->setUser($user);
         $expenseCategory->setAdded($now);
         $expenseCategory->setUpdated($now);
 
@@ -112,7 +154,12 @@ class ExpenseCategoryController extends Controller
         $em->persist($expenseCategory);
         $em->flush();
 
-        $response = $this->serializer->serialize(['data' => $expenseCategory], 'json');
+        $response = $this->serializer->serialize([
+            'data' => $expenseCategory],
+            'json',
+            SerializationContext::create()
+                ->setGroups(['Default'])
+            );
 
         return new JsonResponse($response, 201, [], true);
     }
@@ -139,6 +186,11 @@ class ExpenseCategoryController extends Controller
             throw new HttpException(400, 'You must provide category property');
         }
 
+        $user = $this->authenticator->getUser(
+            $this->authenticator->getCredentials($request),
+            $this->userProvider
+        );
+
         $em = $this->getDoctrine()->getManager();
 
         $category = $em->getRepository(ExpensesCategories::class)
@@ -146,6 +198,10 @@ class ExpenseCategoryController extends Controller
 
         if(!$category) {
             throw new HttpException(404, 'Resource not found.');
+        }
+
+        if($user->getId() != $category->getUser()->getId()) {
+            throw new HttpException(400, 'You do not have permission to edit this resource.');
         }
 
         $category->setCategory($data->getCategory());
@@ -156,7 +212,12 @@ class ExpenseCategoryController extends Controller
 
         $em->flush();
 
-        $category = $this->serializer->serialize(['success' => true, 'data' => $category], 'json');
+        $category = $this->serializer->serialize([
+            'data' => $category],
+            'json',
+            SerializationContext::create()
+                ->setGroups(['Default'])
+        );
 
         return new JsonResponse($category, 200, [], true);
     }
@@ -164,11 +225,12 @@ class ExpenseCategoryController extends Controller
 	/**
 	 * @Route("/expense-categories/{id}", name="delete_expense_category")
      * @Method("DELETE")
+     * @param Request $request
      * @param string $id
      * @return JsonResponse
      * @throws HttpException
 	 */
-    public function delete(string $id): JsonResponse
+    public function delete(Request $request, string $id): JsonResponse
     {
         $em = $this->getDoctrine()->getManager();
         $category = $em->getRepository(ExpensesCategories::class)
@@ -178,10 +240,19 @@ class ExpenseCategoryController extends Controller
             throw new HttpException(404, 'Resource not found.');
         }
 
+        $user = $this->authenticator->getUser(
+            $this->authenticator->getCredentials($request),
+            $this->userProvider
+        );
+
+        if($user->getId() != $category->getUser()->getId()) {
+            throw new HttpException(400, 'You do not have permission to edit this resource.');
+        }
+
         $em->remove($category);
         $em->flush();
 
-        return new JsonResponse(['success' => true]);
+        return new JsonResponse(null, 204);
     }
 }
